Extract stable receiver getter classifier
