+++ src/ILInspector.Analysis/LibraryBodyAnalysisBuilder.cs
@@ -1,4 +1,3 @@
-using System.Collections.Concurrent;
 using System.Collections.Immutable;
 using System.Reflection;
 using System.Reflection.Metadata;
@@ -6,7 +5,6 @@ using System.Reflection.Metadata.Ecma335;
 using System.Reflection.PortableExecutable;
 
 using ILInspector.Findings;
-using ILInspector.Instructions;
 using ILInspector.Metadata;
 
 namespace ILInspector.Analysis;
@@ -27,6 +25,8 @@ internal sealed partial class LibraryBodyAnalysisBuilder :
     readonly PEReader _peReader;
     readonly LibraryBodyPrimaryMetadataResolver
         _primaryMetadataResolver;
+    readonly LibraryBodyStableReceiverGetterClassifier
+        _stableReceiverGetterClassifier;
     readonly LibraryBodyMethodReferenceResolver
         _methodReferenceResolver;
     readonly LibraryBodyAsyncSourceResolver
@@ -50,13 +50,8 @@ internal sealed partial class LibraryBodyAnalysisBuilder :
     readonly string _assemblyName;
     readonly Guid _mvid;
     readonly bool _memorySafetyRulesEnabled;
-    readonly Action<MethodDefinitionHandle>? _stableReceiverGetterClassified;
     readonly Action<TypeDefinitionHandle>? _sourceGeneratedTypeClassified;
     readonly Action? _parallelBuildStarting;
-    readonly ConcurrentDictionary<
-        MethodDefinitionHandle,
-        Lazy<bool>>
-        _stableReceiverGetters = new();
     readonly Dictionary<TypeDefinitionHandle, bool>
         _sourceGeneratedTypes = new();
 
@@ -93,8 +88,6 @@ internal sealed partial class LibraryBodyAnalysisBuilder :
                 null,
                 null);
         _memorySafetyRulesEnabled = DetectMemorySafetyRules();
-        _stableReceiverGetterClassified =
-            stableReceiverGetterClassified;
         _sourceGeneratedTypeClassified =
             sourceGeneratedTypeClassified;
         _parallelBuildStarting = parallelBuildStarting;
@@ -102,6 +95,11 @@ internal sealed partial class LibraryBodyAnalysisBuilder :
             new LibraryBodyMethodReferenceResolver(
                 reader,
                 methodReferenceResolved);
+        _stableReceiverGetterClassifier =
+            new LibraryBodyStableReceiverGetterClassifier(
+                reader,
+                peReader,
+                stableReceiverGetterClassified);
         _primaryMetadataResolver =
             new LibraryBodyPrimaryMetadataResolver(
                 reader,
@@ -109,7 +107,8 @@ internal sealed partial class LibraryBodyAnalysisBuilder :
                 _mvid,
                 _methodReferenceResolver.ResolveMethod,
                 GenericParameterCanBeValueType,
-                IsStableReceiverGetter,
+                _stableReceiverGetterClassifier
+                    .IsStableReceiverGetter,
                 asyncStateMachineTypesBuilt);
         _asyncSourceResolver =
             new LibraryBodyAsyncSourceResolver(
@@ -746,97 +745,6 @@ internal sealed partial class LibraryBodyAnalysisBuilder :
         {
             return false;
         }
-    }
-
-    bool IsStableReceiverGetter(DecodedInstruction instruction)
-    {
-        try
-        {
-            EntityHandle methodHandle = MetadataTokens.EntityHandle(
-                MethodInstructionFacts.OperandInt32(instruction));
-            if (methodHandle.Kind != HandleKind.MethodDefinition)
-                return false;
-
-            var definitionHandle =
-                (MethodDefinitionHandle)methodHandle;
-            var method = _reader.GetMethodDefinition(definitionHandle);
-            bool overridableVirtualCall = instruction.OpCode == ILOpCode.Callvirt
-                && (method.Attributes & MethodAttributes.Virtual) != 0
-                && (method.Attributes & MethodAttributes.Final) == 0
-                && (_reader.GetTypeDefinition(method.GetDeclaringType()).Attributes
-                    & TypeAttributes.Sealed) == 0;
-            if (method.RelativeVirtualAddress == 0
-                || overridableVirtualCall
-                || !_reader.GetString(method.Name).StartsWith(
-                    "get_",
-                    StringComparison.Ordinal))
-            {
-                return false;
-            }
-
-            return _stableReceiverGetters.GetOrAdd(
-                definitionHandle,
-                handle => new Lazy<bool>(
-                    () => ClassifyStableReceiverGetter(handle),
-                    LazyThreadSafetyMode.ExecutionAndPublication)).Value;
-        }
-        catch (Exception ex) when (ex is BadImageFormatException
-            or InvalidOperationException
-            or ArgumentException
-            or OverflowException
-            or InvalidCastException)
-        {
-            return false;
-        }
-    }
-
-    bool ClassifyStableReceiverGetter(
-        MethodDefinitionHandle methodHandle)
-    {
-        _stableReceiverGetterClassified?.Invoke(methodHandle);
-        MethodDefinition method =
-            _reader.GetMethodDefinition(methodHandle);
-        var body = _peReader.GetMethodBody(method.RelativeVirtualAddress);
-        if (body.ExceptionRegions.Length != 0)
-            return false;
-        DecodedInstruction? first = null;
-        DecodedInstruction? fieldLoad = null;
-        DecodedInstruction? third = null;
-        int count = 0;
-        foreach (DecodedInstruction instruction
-            in InstructionDecoder.Decode(body.GetILBytes() ?? []))
-        {
-            if (instruction.OpCode == ILOpCode.Nop)
-                continue;
-            switch (count++)
-            {
-                case 0:
-                    first = instruction;
-                    break;
-                case 1:
-                    fieldLoad = instruction;
-                    break;
-                case 2:
-                    third = instruction;
-                    break;
-                default:
-                    return false;
-            }
-        }
-        if (count != 3
-            || first is not { OpCode: ILOpCode.Ldarg_0 }
-            || fieldLoad is not { OpCode: ILOpCode.Ldfld }
-            || third is not { OpCode: ILOpCode.Ret })
-        {
-            return false;
-        }
-
-        EntityHandle fieldHandle = MetadataTokens.EntityHandle(
-            MethodInstructionFacts.OperandInt32(fieldLoad));
-        return fieldHandle.Kind == HandleKind.FieldDefinition
-            && (_reader.GetFieldDefinition(
-                    (FieldDefinitionHandle)fieldHandle).Attributes
-                & FieldAttributes.InitOnly) != 0;
     }
 
     bool ConstraintCanIncludeValueType(EntityHandle constraint)
